vision/requirements.txt
torch==2.0.1
torchvision==0.15.2
torchdiffeq==0.2.3

pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==0.9.7

tensorboard
opencv-python==4.8.0.76
augly

jupyterlab
seaborn
pandas
path
easydict
scipy
matplotlib
numpy
tqdm
PyYAML

onnx==1.14.0
onnxruntime==1.15.1

pytest

labelbox==3.55.0
xmltodict==0.13.0
boto3==1.29.1
pyheif
